ImportSubscriptionsAdapter in now renamed as MultiSelectListPreferenceAdapter.  This adapter is used by MultiSelectListPreferenceDialog - a common class that lists items, allows to user to select multiple items and can select/deselect all items.  Class used by SubscriptionsManager and preferred languages preference.  As a result, the Preferred Language(s) preference now allows the users to easily choose the language(s) they fancy.

+++ app/src/main/java/free/rm/skytube/gui/fragments/preferences/VideoBlockerPreferenceFragment.java
@@ -1,13 +1,29 @@
 package free.rm.skytube.gui.fragments.preferences;
 
+import android.content.Context;
+import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.preference.MultiSelectListPreference;
 import android.preference.Preference;
 import android.preference.PreferenceFragment;
+import android.support.annotation.NonNull;
 import android.widget.Toast;
+
+import com.afollestad.materialdialogs.DialogAction;
+import com.afollestad.materialdialogs.MaterialDialog;
+
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+
+import free.rm.skytube.BuildConfig;
 import free.rm.skytube.R;
 import free.rm.skytube.businessobjects.db.BlockedChannelsDb;
+import free.rm.skytube.gui.businessobjects.MultiSelectListPreferenceDialog;
+import free.rm.skytube.gui.businessobjects.MultiSelectListPreferenceItem;
 
 /**
  * Video blocker preference.
@@ -42,22 +58,36 @@ public class VideoBlockerPreferenceFragment extends PreferenceFragment {
 			public boolean onPreferenceChange(Preference preference, Object o) {
 				// The selected channels are the channels that user wishes to unblock.
 				// Object o gives us a set of blocked channel names.
-				Collection<String> selectedChannels = (Collection<String>) o;
+				final Collection<String> selectedChannels = (Collection<String>) o;
 
-				for (String channel : selectedChannels) {
-					BlockedChannelsDb.getBlockedChannelsDb().remove(channel);
+				if (!selectedChannels.isEmpty()) {
+					for (String channel : selectedChannels) {
+						BlockedChannelsDb.getBlockedChannelsDb().remove(channel);
+					}
+
+					//We set the last updated version of DB here again
+					//So we will not see unblocked channels.
+					String[] channels = getBlockedChannels();
+					channelBlacklistPreference.setEntryValues(channels);
+					channelBlacklistPreference.setEntries(channels);
+
+					Toast.makeText(getActivity(), R.string.channel_blacklist_updated, Toast.LENGTH_LONG).show();
 				}
-
-				//We set the last updated version of DB here again
-				//So we will not see unblocked channels.
-				String[] channels = getBlockedChannels();
-				channelBlacklistPreference.setEntryValues(channels);
-				channelBlacklistPreference.setEntries(channels);
 				
 				return true;
 			}
 		});
+
+		final Preference preferredLanguagesPreference = findPreference(getString(R.string.pref_key_preferred_languages));
+		preferredLanguagesPreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+			@Override
+			public boolean onPreferenceClick(Preference preference) {
+				displayPreferredLanguageDialog();
+				return true;
+			}
+		});
 	}
+
 
 	/**
 	 * Method to get updated version of blocked channels
@@ -69,4 +99,97 @@ public class VideoBlockerPreferenceFragment extends PreferenceFragment {
 		return blockedChannelsDb.getBlockedChannelsListName().
 				toArray(new String[blockedChannelsDb.getBlockedChannelsListName().size()]);
 	}
+
+
+	/**
+	 * Display a dialog which allows the user to select his preferred language(s).
+	 */
+	private void displayPreferredLanguageDialog() {
+		final MultiSelectListPreferenceDialog prefLangDialog = new MultiSelectListPreferenceDialog(getActivity(), getLanguagesAvailable());
+		prefLangDialog.title(R.string.pref_title_preferred_languages)
+				.positiveText(R.string.ok)
+				.onPositive(new MaterialDialog.SingleButtonCallback() {
+					@Override
+					public void onClick(@NonNull MaterialDialog dialog, @NonNull DialogAction which) {
+						final Set<String> preferredLangIsoCodes = prefLangDialog.getSelectedItemsIds();
+
+						// if at least one language was selected, then save the settings
+						if (preferredLangIsoCodes.size() > 0) {
+							SharedPreferences sharedPref = getActivity().getPreferences(Context.MODE_PRIVATE);
+							SharedPreferences.Editor editor = sharedPref.edit();
+							editor.putStringSet(getString(R.string.pref_key_preferred_languages), prefLangDialog.getSelectedItemsIds());
+							editor.apply();
+
+							Toast.makeText(getActivity(), R.string.preferred_lang_updated, Toast.LENGTH_LONG).show();
+						} else {
+							// no languages were selected... action is ignored
+							Toast.makeText(getActivity(), R.string.no_preferred_lang_selected, Toast.LENGTH_LONG).show();
+						}
+					}
+				})
+				.negativeText(R.string.cancel)
+				.onNegative(new MaterialDialog.SingleButtonCallback() {
+					@Override
+					public void onClick(@NonNull MaterialDialog dialog, @NonNull DialogAction which) {
+						dialog.dismiss();
+					}
+				})
+				.build()
+				.show();
+	}
+
+
+	/**
+	 * @return A list of languages supported by YouTube - if a language is currently preferred by the
+	 * user, the language is ticked/marked.
+	 */
+	private List<MultiSelectListPreferenceItem> getLanguagesAvailable() {
+		List<MultiSelectListPreferenceItem> languagesAvailable = getLanguagesList();
+		Set<String>                         preferredLanguages = getPreferredLanguages();
+
+		for (MultiSelectListPreferenceItem language : languagesAvailable) {
+			language.isChecked = preferredLanguages.contains(language.id);
+		}
+
+		return languagesAvailable;
+	}
+
+
+	/**
+	 * @return A list of languages supported by YouTube.
+	 */
+	private List<MultiSelectListPreferenceItem> getLanguagesList() {
+		String[] languagesNames = getResources().getStringArray(R.array.languages_names);
+		String[] languagesIsoCodes = getResources().getStringArray(R.array.languages_iso639_codes);
+		List<MultiSelectListPreferenceItem> languageAvailableList = new ArrayList<>();
+
+		if (BuildConfig.DEBUG  &&  (languagesNames.length != languagesIsoCodes.length)) {
+			throw new AssertionError("languages names array is NOT EQUAL to languages ISO codes array.");
+		}
+
+		for (int i = 0;  i < languagesNames.length;  i++) {
+			languageAvailableList.add(new MultiSelectListPreferenceItem(languagesIsoCodes[i], languagesNames[i]));
+		}
+
+		return languageAvailableList;
+	}
+
+
+	/**
+	 * @return A set of languages preferred by the user.
+	 */
+	private Set<String> getPreferredLanguages() {
+		SharedPreferences pref = getActivity().getPreferences(Context.MODE_PRIVATE);
+		return pref.getStringSet(getString(R.string.pref_key_preferred_languages), getDefaultPreferredLanguages());
+	}
+
+
+	/**
+	 * @return The default setting for preferred languages (i.e. no language preference).
+	 */
+	private Set<String> getDefaultPreferredLanguages() {
+		String[] languagesIsoCodes = getResources().getStringArray(R.array.languages_iso639_codes);
+		return new HashSet<>(Arrays.asList(languagesIsoCodes));
+	}
+
 }
